+++ src/pages/celublog/CelublogStyle.scss
@@ -370,11 +370,11 @@
 }
 
 .celub-rulemake-input1{
-  width: 60%;
+  width: 55%;
   margin-right: 0.5rem;
 }
 .celub-rulemake-input2{
-  width: 30%;
+  width: 45%;
 }
 
 .celub-addrule-btn{
